+++ src/math/Vector3.java
@@ -9,7 +9,7 @@ package math;
  * @author Alexey
  */
 public class Vector3 {
-    private float[] values; /*Значения хранятся в виде массива из трёх элементов*/
+    private double[] values; /*Значения хранятся в виде массива из трёх элементов*/
     
     /**
      * Создаёт экземпляр вектора на основе значений трёх составляющих
@@ -17,15 +17,15 @@ public class Vector3 {
      * @param y вторая составляющая, описывающая Y-координату
      * @param z третья составляющая, описывающая Z-координату
      */
-    public Vector3(float x, float y, float z) {
-        values = new float[]{x, y, z};
+    public Vector3(double x, double y, double z) {
+        values = new double[]{x, y, z};
     }
 
     /**
      * X-составляющая вектора
      * @return X-составляющая вектора
      */
-    public float getX() {
+    public double getX() {
         return values[0];
     }
 
@@ -33,7 +33,7 @@ public class Vector3 {
      * Y-составляющая вектора
      * @return Y-составляющая вектора
      */
-    public float getY() {
+    public double getY() {
         return values[1];
     }
 
@@ -41,7 +41,7 @@ public class Vector3 {
      * Z-составляющая вектора
      * @return Z-составляющая вектора
      */
-    public float getZ() {
+    public double getZ() {
         return values[2];
     }
     
@@ -50,20 +50,20 @@ public class Vector3 {
      * @param idx порядковый номер
      * @return значение составляющей вектора
      */
-    public float at(int idx) {
+    public double at(int idx) {
         return values[idx];
     }
     
-    private static final float EPSILON = 1e-10f;
+    private static final double EPSILON = 1e-10f;
     /**
      * Метод, возвращающий длину вектора
      * @return длина вектора
      */
-    public float length() {
-        float lenSqr = values[0] * values[0] + values[1] * values[1] + values[2] * values[2];
+    public double length() {
+        double lenSqr = values[0] * values[0] + values[1] * values[1] + values[2] * values[2];
         if (lenSqr < EPSILON)
             return 0;
-        return (float)Math.sqrt(lenSqr);
+        return (double)Math.sqrt(lenSqr);
     }
     
 }
